+++ app/items/[uuid]/page.tsx
@@ -3,7 +3,7 @@ import { Metadata } from "next";
 import PageLayout from "../../../src/components/pageLayout/pageLayout";
 import Item from "../../../src/components/items/item";
 import { getItemData } from "../../../src/utils/utils";
-import { ItemModel } from "../../../src/models/item";
+import { ItemModel } from "src/models/item";
 
 type ItemProps = {
   params: {