+++ app/page.tsx
@@ -1,4 +1,15 @@
 import HomePage from "./src/components/pages/homePage/homePage";
-export default function Home() {
-  return <HomePage />;
+import { getHomePageData, getFeaturedItemData } from "@/src/utils/api";
+import { FeaturedItemDataType } from "../app/src/types/FeaturedItemDataType";
+import { revalidatePath } from "next/cache";
+
+export default async function Home() {
+  revalidatePath("/");
+  const swimLaneData = await getHomePageData();
+  const featuredItemData: FeaturedItemDataType = await getFeaturedItemData();
+  return (
+    <HomePage
+      data={{ swimLaneData: swimLaneData, featuredItemData: featuredItemData }}
+    />
+  );
 }