+++ CHANGELOG.md
@@ -7,11 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Rendered division landing page (DR-3093)
+- Created item and item card models and mocks (DR-3094)
+
 ## [0.1.11] 2024-08-05
 
 ### Added
 
 - Added Github Actions deployment workflows, Travis deployment script removed (DR-3019)
+
 
 ## [0.1.10] 2024-08-01