+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Rendered collections page with preliminary search/sort (DR-3099)
 - Rendered division landing page (DR-3093)
 - Created item and item card models and mocks (DR-3094)
 
@@ -17,7 +18,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added Github Actions deployment workflows, Travis deployment script removed (DR-3019)
-
 
 ## [0.1.10] 2024-08-01