+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.1.12] 2024-08-05
+
+### Updated
+- Updated swim lanes and featured item features to use server side api calls (DR-3016)
+
+### Added
+- Update middleware file to not redirect urls created for phase 2-4 in 'qa' environment (DR-3122)
+- Rendered `/collections` page with preliminary search/sort (DR-3099)
+- Rendered `/collections/lane/[slug]` collection lane landing page (DR-3106)
+- Rendered `/divisions/[slug]` division landing page (DR-3093)
+- Created item and item card models and mocks (DR-3094)
+- Added `/api/changelog` endpoint (DR-3048)
+
 ## [0.1.11] 2024-08-05
 
 ### Added